moved AssetsList to separate file
