+++ src/components/WatchlistEdit/WatchlistEdit.module.scss
@@ -29,39 +29,6 @@
   margin: 16px 0 8px;
 }
 
-.wrapperList {
-  position: relative;
-
-  &::after {
-    content: '';
-    position: absolute;
-    bottom: -3px;
-    left: 0;
-    right: 0;
-    height: 12px;
-    background: linear-gradient(180deg, rgba(255, 255, 255, 0), var(--white));
-  }
-}
-
-.project {
-  @include text-normal;
-
-  font-size: 14px;
-  line-height: 20px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.name {
-  margin-right: 5px;
-  color: var(--mirage);
-}
-
-.actionBtn {
-  margin-right: 10px;
-}
-
 .actions {
   padding: 10px 15px;
 }
